# Conflicts:
#	src/main/java/uhh_lt/webserver/SolrConnectTester.java

+++ src/test/java/uhh_lt.webserver/MieterClassifierTest.java
@@ -1,9 +1,8 @@
 package uhh_lt.webserver;
 
 import org.junit.Test;
-import uhh_lt.classifier.MieterClassifier;
 
-import static org.junit.Assert.assertTrue;
+import static org.junit.Assert.*;
 
 public class MieterClassifierTest {
 
@@ -17,23 +16,23 @@ public class MieterClassifierTest {
     @Test
     public void classifySimple()
     {
-        Double mC = mieterClassifier.classify("Ich bin Mieter");
+        float mC = mieterClassifier.classify("Ich bin Mieter");
         assertTrue("Die Mieterwahrscheinlichkeit muss 1 sein", 1.0 == mC);
-        Double vC = mieterClassifier.classify("Ich bin Vermieter");
+        float vC = mieterClassifier.classify("Ich bin Vermieter");
         assertTrue("Die Mieterwahrscheinlichkeit muss 0 sein", 0.0 == vC);
-        Double nC = mieterClassifier.classify("Ich bin weder Mieter noch Vermieter");
+        float nC = mieterClassifier.classify("Ich bin weder Mieter noch Vermieter");
         assertTrue("Die Mieterwahrscheinlichkeit muss 0.5 sein", 0.5 == nC);
     }
 
     @Test
     public void classifyMehrfachnennung()
     {
-        Double mC = mieterClassifier.classify("Ich bin Mieter, bin Vermieter und ich miete eine Wohnung");
+        float mC = mieterClassifier.classify("Ich bin Mieter, bin Vermieter und ich miete eine Wohnung");
         System.out.println(mC);
         assertTrue("Die Mieterwahrscheinlichkeit muss 2/3 sein", ((float) 2/3) == mC);
-        Double nC = mieterClassifier.classify("Ich bin Mieter  und bin Vermieter und ich vermiete und ich miete");
+        float nC = mieterClassifier.classify("Ich bin Mieter  und bin Vermieter und ich vermiete und ich miete");
         assertTrue("Die Mieterwahrscheinlichkeit muss 0.5 sein", 0.5 == nC);
-        Double aC = mieterClassifier.classify("Ich bin Mieter  und bin Mieter und ich vermiete und ich miete");
+        float aC = mieterClassifier.classify("Ich bin Mieter  und bin Mieter und ich vermiete und ich miete");
         assertTrue("Die Mieterwahrscheinlichkeit muss 0.75 sein", 0.75 == aC);
     }
 }
